Struct aws_sdk_qldb::client::fluent_builders::StreamJournalToKinesis

pub struct StreamJournalToKinesis<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> StreamJournalToKinesis<C>

pub async fn send(
self
) -> Result<StreamJournalToKinesisOutput, SdkError<StreamJournalToKinesisError>> where
C: SmithyConnector,

pub fn ledger_name(self, input: impl Into<String>) -> Self

The name of the ledger.

pub fn set_ledger_name(self, input: Option<String>) -> Self

pub fn role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal stream to write data records to a Kinesis Data Streams resource.

pub fn set_role_arn(self, input: Option<String>) -> Self

pub fn tags(self, k: impl Into<String>, v: impl Into<Option<String>>) -> Self

The key-value pairs to add as tags to the stream that you want to create. Tag keys are case sensitive. Tag values are case sensitive and can be null.

pub fn set_tags(self, input: Option<HashMap<String, Option<String>>>) -> Self

pub fn inclusive_start_time(self, input: Instant) -> Self

The inclusive start date and time from which to start streaming journal data. This parameter must be in ISO 8601 date and time format and in Universal Coordinated Time (UTC). For example: 2019-06-13T21:36:34Z.

The InclusiveStartTime cannot be in the future and must be before ExclusiveEndTime.

If you provide an InclusiveStartTime that is before the ledger's CreationDateTime, QLDB effectively defaults it to the ledger's CreationDateTime.

pub fn set_inclusive_start_time(self, input: Option<Instant>) -> Self

pub fn exclusive_end_time(self, input: Instant) -> Self

The exclusive date and time that specifies when the stream ends. If you don't define this parameter, the stream runs indefinitely until you cancel it.

The ExclusiveEndTime must be in ISO 8601 date and time format and in Universal Coordinated Time (UTC). For example: 2019-06-13T21:36:34Z.

pub fn set_exclusive_end_time(self, input: Option<Instant>) -> Self

pub fn kinesis_configuration(self, input: KinesisConfiguration) -> Self

The configuration settings of the Kinesis Data Streams destination for your stream request.

pub fn set_kinesis_configuration(
self,
input: Option<KinesisConfiguration>
) -> Self

pub fn stream_name(self, input: impl Into<String>) -> Self

The name that you want to assign to the QLDB journal stream. User-defined names can help identify and indicate the purpose of a stream.

Your stream name must be unique among other active streams for a given ledger. Stream names have the same naming constraints as ledger names, as defined in Quotas in Amazon QLDB in the Amazon QLDB Developer Guide.

pub fn set_stream_name(self, input: Option<String>) -> Self