Module aws_sdk_qldb::model

Modules

journal_kinesis_stream_description

See JournalKinesisStreamDescription

journal_s3_export_description

See JournalS3ExportDescription

kinesis_configuration

See KinesisConfiguration

ledger_summary

See LedgerSummary

s3_encryption_configuration

See S3EncryptionConfiguration

s3_export_configuration

See S3ExportConfiguration

value_holder

See ValueHolder

Structs

JournalKinesisStreamDescription

Information about an Amazon QLDB journal stream, including the Amazon Resource Name

JournalS3ExportDescription

Information about a journal export job, including the ledger name, export ID, creation

KinesisConfiguration

The configuration settings of the Amazon Kinesis Data Streams destination for an Amazon QLDB journal

LedgerSummary

Information about a ledger, including its name, state, and when it was created.

S3EncryptionConfiguration

The encryption settings that are used by a journal export job to write data in an

S3ExportConfiguration

The Amazon Simple Storage Service (Amazon S3) bucket location in which a journal export job writes the journal

ValueHolder

A structure that can contain a value in multiple encoding formats.

Enums

ErrorCause

ExportStatus

LedgerState

PermissionsMode

S3ObjectEncryptionType

StreamStatus