Module aws_sdk_qldb::operation

Structs

CancelJournalKinesisStream

Ends a given Amazon QLDB journal stream. Before a stream can be canceled, its current

CreateLedger

Creates a new ledger in your AWS account in the current Region.

DeleteLedger

Deletes a ledger and all of its contents. This action is irreversible.

DescribeJournalKinesisStream

Returns detailed information about a given Amazon QLDB journal stream. The output

DescribeJournalS3Export

Returns information about a journal export job, including the ledger name, export ID,

DescribeLedger

Returns information about a ledger, including its state and when it was created.

ExportJournalToS3

Exports journal contents within a date and time range from a ledger into a specified

GetBlock

Returns a block object at a specified address in a journal. Also returns a proof of the

GetDigest

Returns the digest of a ledger at the latest committed block in the journal. The

GetRevision

Returns a revision data object for a specified document ID and block address. Also

ListJournalKinesisStreamsForLedger

Returns an array of all Amazon QLDB journal stream descriptors for a given ledger. The

ListJournalS3Exports

Returns an array of journal export job descriptions for all ledgers that are associated

ListJournalS3ExportsForLedger

Returns an array of journal export job descriptions for a specified ledger.

ListLedgers

Returns an array of ledger summaries that are associated with the current AWS account

ListTagsForResource

Returns all tags for a specified Amazon QLDB resource.

StreamJournalToKinesis

Creates a journal stream for a given Amazon QLDB ledger. The stream captures every

TagResource

Adds one or more tags to a specified Amazon QLDB resource.

UntagResource

Removes one or more tags from a specified Amazon QLDB resource. You can specify up to 50

UpdateLedger

Updates properties on a ledger.

UpdateLedgerPermissionsMode

Updates the permissions mode of a ledger.